+++ src/build.task.ts
@@ -8,10 +8,7 @@ import {load_config} from './config/config.js';
 import type {Gro_Config} from './config/config.js';
 import type {Task_Events as Server_Task_Events} from './server.task.js';
 import type {Adapter_Context, Adapter} from './adapt/adapter.js';
-import {build_source_directory} from './build/build_source_directory.js';
-import {generate_types} from './build/ts_build_helpers.js';
-import {paths, to_types_build_dir} from './paths.js';
-import {clean} from './fs/clean.js';
+import {build_source} from './build/build_source.js';
 
 export interface Task_Args extends Args {
 	map_input_options?: Map_Input_Options;
@@ -21,12 +18,11 @@ export interface Task_Args extends Args {
 
 export interface Task_Events extends Server_Task_Events {
 	'build.create_config': (config: Gro_Config) => void;
-	'build.build_types': void;
 	'build.build_src': void;
 }
 
 export const task: Task<Task_Args, Task_Events> = {
-	description: 'build the project',
+	summary: 'build the project',
 	dev: false,
 	run: async (ctx): Promise<void> => {
 		const {fs, dev, log, events} = ctx;
@@ -36,15 +32,6 @@ export const task: Task<Task_Args, Task_Events> = {
 
 		const timings = new Timings(); // TODO belongs in ctx
 
-		await clean(fs, {build_prod: true}, log);
-
-		// Build all types so they're available.
-		// TODO refactor? maybe lazily build types only when a builder wants them
-		const timing_to_build_types = timings.start('build_types');
-		await generate_types(paths.source, to_types_build_dir(), true);
-		timing_to_build_types();
-		events.emit('build.build_types');
-
 		const timing_to_load_config = timings.start('load config');
 		const config = await load_config(fs, dev);
 		timing_to_load_config();
@@ -53,7 +40,7 @@ export const task: Task<Task_Args, Task_Events> = {
 		// Build everything with esbuild and Gro's `Filer` first.
 		// These production artifacts are then available to all adapters.
 		const timing_to_build_src = timings.start('build_src');
-		await build_source_directory(fs, config, dev, log);
+		await build_source(fs, config, dev, log);
 		timing_to_build_src();
 		events.emit('build.build_src');