+++ src/config/config.ts
@@ -11,18 +11,14 @@ import type {Assignable, Result} from '@feltcoop/felt/util/types.js';
 import {to_array} from '@feltcoop/felt/util/array.js';
 
 import {paths, to_build_out_path, CONFIG_BUILD_PATH, DIST_DIRNAME} from '../paths.js';
-import {
-	is_system_build_config,
-	is_config_build_config,
-	normalize_build_configs,
-	validate_build_configs,
-} from '../build/build_config.js';
+import {normalize_build_configs, validate_build_configs} from '../build/build_config.js';
 import type {Adapt_Builds} from '../adapt/adapter.js';
 import type {Build_Config, Build_Config_Partial} from '../build/build_config.js';
 import {
 	DEFAULT_ECMA_SCRIPT_TARGET,
 	NODE_LIBRARY_BUILD_NAME,
 	CONFIG_BUILD_CONFIG,
+	SYSTEM_BUILD_NAME,
 } from '../build/default_build_config.js';
 import type {Ecma_Script_Target} from '../build/ts_build_helpers.js';
 import type {Served_Dir_Partial} from '../build/served_dir.js';
@@ -59,7 +55,6 @@ export interface Gro_Config {
 	readonly port: number;
 	readonly log_level: Log_Level;
 	readonly serve?: Served_Dir_Partial[];
-	readonly config_build_config: Build_Config;
 	readonly system_build_config: Build_Config;
 	readonly primary_browser_build_config: Build_Config | null; // TODO improve this, too rigid
 }
@@ -154,13 +149,8 @@ export const load_config = async (
 	const {config_source_id} = paths;
 	let config: Gro_Config;
 	if (await fs.exists(config_source_id)) {
-		const {build_source_directory} = await import('../build/build_source_directory.js');
-		const bootstrap_config = await to_config(
-			{builds: [CONFIG_BUILD_CONFIG], sourcemap: dev},
-			options,
-			'gro/build/default_build_config.ts',
-		);
-		await build_source_directory(fs, bootstrap_config, dev, log);
+		const {build_source} = await import('../build/build_source.js');
+		await build_source(fs, to_bootstrap_config(), dev, log, false);
 
 		// The project has a `gro.config.ts`, so import it.
 		// If it's not already built, we need to bootstrap the config and use it to compile everything.
@@ -197,14 +187,29 @@ export const to_config = async (
 
 	const extended_config = base_config ? {...base_config, ...config_partial} : config_partial;
 
-	const config = normalize_config(extended_config);
+	const config = normalize_config(extended_config, options.dev);
 
-	const validate_result = await validate_config(options.fs, config);
+	const validate_result = await validate_config(options.fs, config, options.dev);
 	if (!validate_result.ok) {
 		throw Error(`Invalid Gro config at '${path}': ${validate_result.reason}`);
 	}
 
 	return config;
+};
+
+const to_bootstrap_config = (): Gro_Config => {
+	return {
+		sourcemap: false, // TODO or always true?
+		host: DEFAULT_SERVER_HOST,
+		port: DEFAULT_SERVER_PORT,
+		log_level: DEFAULT_LOG_LEVEL,
+		adapt: () => null,
+		builds: [CONFIG_BUILD_CONFIG],
+		publish: null,
+		target: DEFAULT_ECMA_SCRIPT_TARGET,
+		system_build_config: null!,
+		primary_browser_build_config: null,
+	};
 };
 
 const validate_config_module = (config_module: any): Result<{}, {reason: string}> => {
@@ -217,16 +222,17 @@ const validate_config_module = (config_module: any): Result<{}, {reason: string}
 const validate_config = async (
 	fs: Filesystem,
 	config: Gro_Config,
+	dev: boolean,
 ): Promise<Result<{}, {reason: string}>> => {
-	const build_configs_result = await validate_build_configs(fs, config.builds);
+	const build_configs_result = await validate_build_configs(fs, config.builds, dev);
 	if (!build_configs_result.ok) return build_configs_result;
 	return {ok: true};
 };
 
-const normalize_config = (config: Gro_Config_Partial): Gro_Config => {
-	const build_configs = normalize_build_configs(to_array(config.builds || null));
+const normalize_config = (config: Gro_Config_Partial, dev: boolean): Gro_Config => {
+	const build_configs = normalize_build_configs(to_array(config.builds || null), dev);
 	return {
-		sourcemap: process.env.NODE_ENV !== 'production', // TODO maybe default to tsconfig?
+		sourcemap: dev, // TODO maybe default to tsconfig?
 		host: DEFAULT_SERVER_HOST,
 		port: DEFAULT_SERVER_PORT,
 		log_level: DEFAULT_LOG_LEVEL,
@@ -238,8 +244,7 @@ const normalize_config = (config: Gro_Config_Partial): Gro_Config => {
 				? config.publish
 				: to_default_publish_dirs(build_configs),
 		target: config.target || DEFAULT_ECMA_SCRIPT_TARGET,
-		config_build_config: build_configs.find((b) => is_config_build_config(b))!,
-		system_build_config: build_configs.find((b) => is_system_build_config(b))!,
+		system_build_config: build_configs.find((b) => b.name === SYSTEM_BUILD_NAME)!,
 		// TODO instead of `primary` build configs, we want to be able to mount any number of them at once,
 		// so this is a temp hack that just chooses the first browser build
 		primary_browser_build_config: build_configs.find((b) => b.platform === 'browser') || null,