+++ src/task/fixtures/test_task_module.task_fixture.ts
@@ -1,7 +1,7 @@
 import type {Task} from '../task.js';
 
 export const task: Task = {
-	description: 'a test task for basic task behavior',
+	summary: 'a test task for basic task behavior',
 	run: async ({log, args}) => {
 		log.info('test task 1!', args);
 		return args;