+++ src/typecheck.task.ts
@@ -4,7 +4,7 @@ import type {Task} from './task/task.js';
 import {Task_Error} from './task/task.js';
 
 export const task: Task = {
-	description: 'typecheck the project without emitting any files',
+	summary: 'typecheck the project without emitting any files',
 	run: async () => {
 		const typecheck_result = await spawn_process('npx', ['tsc', '--noEmit']);
 		if (!typecheck_result.ok) {